+++ composables/transaction/types.ts
@@ -1,0 +1,34 @@
+import { Interaction } from '@kodadot1/minimark'
+import { ShoppingActions } from '@/utils/shoppingActions'
+
+export type ActionList = {
+  interaction: Interaction.LIST
+  urlPrefix: string
+  price: string
+  nftId: string
+  successMessage?: string
+  errorMessage?: string
+}
+
+export type ActionSend = {
+  interaction: Interaction.SEND
+  urlPrefix: string
+  tokenId: string
+  address: string
+  nftId: string
+  successMessage?: string
+  errorMessage?: string
+}
+
+export type ActionOffer = {
+  interaction: typeof ShoppingActions.MAKE_OFFER
+  urlPrefix: string
+  tokenId: string
+  day: number
+  price: number
+  currentOwner: string
+  successMessage?: string
+  errorMessage?: string
+}
+
+export type Actions = ActionList | ActionSend | ActionOffer
